updated the Jupyter notebook, and moved it to a subdirectory `ipynb`.
Note that if one runs `jupyter notebook` for a notebook that resides
in a directory containing a `Project.toml`,
apparently Julia assumes this project as its focus.
The effect in my installation is that in this situation `Pkg.status()`
shows a wrong version of GAP:
Instead of the develop version Julia uses otherwise, now an outdated
released version is used.

I think this is a very bad behaviour.